# Reduce allocations during image decoding.