--- a/src/Pfim/dds/Dxt5Dds.cs
+++ b/src/Pfim/dds/Dxt5Dds.cs
@@ -7,8 +7,6 @@ namespace Pfim
         private const byte PIXEL_DEPTH = 4;
         private const byte DIV_SIZE = 4;
 
-        private readonly byte[] alpha = new byte[8];
-
         public override int BitsPerPixel => 8 * PIXEL_DEPTH;
         public override ImageFormat Format => ImageFormat.Rgba32;
         protected override byte DivSize => DIV_SIZE;
@@ -20,8 +18,10 @@ namespace Pfim
 
         protected override byte PixelDepthBytes => PIXEL_DEPTH;
 
-        protected override int Decode(byte[] stream, byte[] data, int streamIndex, uint dataIndex, uint stride)
+        protected override unsafe int Decode(byte[] stream, byte[] data, int streamIndex, uint dataIndex, uint stride)
         {
+            byte* alpha = stackalloc byte[8];
+
             streamIndex = Bc5Dds.ExtractGradient(alpha, stream, streamIndex);
 
             ulong alphaCodes = stream[streamIndex++];
@@ -43,7 +43,7 @@ namespace Pfim
             var c1 = ColorFloatRgb.FromRgb565(color1);
 
             (var i0, var i1) = (c0.As8Bit(), c1.As8Bit());
-            Color888[] colors = new[] { i0, i1, c0.Lerp(c1, 1f / 3).As8Bit(), c0.Lerp(c1, 2f / 3).As8Bit() };
+            Color888* colors = stackalloc Color888[] { i0, i1, c0.Lerp(c1, 1f / 3).As8Bit(), c0.Lerp(c1, 2f / 3).As8Bit() };
 
             for (int alphaShift = 0; alphaShift < 48; alphaShift += 12)
             {