@nickbabcock 2024-08-23
Nicely done, the reduction in memory allocations speak for themselves. LGTM, I can look to merge this soon.